State holders refactoring
