+++ composeApp/src/commonMain/kotlin/ui/dialogs/adddevice/AddDeviceViewModel.kt
@@ -1,0 +1,11 @@
+package ui.dialogs.adddevice
+
+import androidx.lifecycle.ViewModel
+import sharedData.Repository
+
+class AddDeviceViewModel (
+    private val repository: Repository
+) : ViewModel() {
+
+
+}
